[clang][test] add TestLanguage.def to specify all tested language versions
Adds a def file to have a single location where tested language versions are
specified. Removes the need to update multiple locations in the testing
infrastructure to add a new language version to be tested.
Test instatiation can now include all languages without needing to
specify them.
This patch also adds pretty printing for instantiated test names.
That means, that a test instantiated with C++23 will have the
name `...TestSuite/TestName/CXX23` instead ending with some number
(index of the argument for instantiation of the test), which improves a
better experience when encountering a test failure with a specific
language version. The suffix will also contain an `_win` if the target
contains `win`.